requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# Pydantic
pydantic>=2.11.0
pydantic-settings>=2.5.2
email-validator>=2.1.0

# Slack
slack-bolt>=1.18.1
slack-sdk>=3.27.1
aiohttp>=3.9.0

# Google Gemini
google-genai>=1.45.0

# Background Tasks
apscheduler>=3.10.4

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=24.0.0
ruff>=0.1.11
